Solve circular dependencies
