+++ apps/auto-posts/src/app/twitter/twitter.service.ts
@@ -28,14 +28,15 @@ import TwitterApi, {
 } from 'twitter-api-v2';
 import { Environment } from '@castcle-api/environments';
 import { Cron, CronExpression } from '@nestjs/schedule';
-import { ContentService, SocialSyncService } from '@castcle-api/database';
 import {
+  ContentService,
   SocialProvider,
-  SocialSyncDocument
-} from '@castcle-api/database/schemas';
+  SocialSyncService
+} from '@castcle-api/database';
 import { CastLogger, CastLoggerOptions } from '@castcle-api/logger';
 import { Author, SaveContentDto } from '@castcle-api/database/dtos';
 import { COMMON_SIZE_CONFIGS, Downloader, Image } from '@castcle-api/utils/aws';
+import { SocialSyncDocument } from '@castcle-api/database/schemas';
 
 @Injectable()
 export class TwitterService {
