+++ libs/database/src/lib/dtos/user.dto.ts
@@ -22,8 +22,7 @@
  */
 import { ApiProperty } from '@nestjs/swagger';
 import { IsBoolean, IsEnum, IsNotEmpty, IsString } from 'class-validator';
-import { SocialProvider } from '../schemas/social-sync.schema';
-import { PageVerified, UserVerified } from '../schemas/user.schema';
+import { SocialProvider, PageVerified, UserVerified } from '../models';
 import { CastcleImage, Pagination } from './common.dto';
 
 class UserImage {
