+++ libs/database/src/lib/schemas/account.schema-factory.ts
@@ -21,5 +21,27 @@
  * or have any questions.
  */
 
-export * from './lib/models';
-export * from './lib/database.module';
+import { Model, Schema } from 'mongoose';
+import { postAccountSave, preAccountSave } from '../hooks/account.save';
+import { AccountDocument, AccountSchema } from './account.schema';
+import { CredentialDocument } from './credential.schema';
+import { UserDocument } from './user.schema';
+
+export const AccountSchemaFactory = (
+  credentialModel: Model<CredentialDocument>,
+  userModel: Model<UserDocument>
+): Schema<any> => {
+  AccountSchema.pre('save', function (next) {
+    preAccountSave(this as AccountDocument);
+    next();
+  });
+  AccountSchema.post('save', async function (doc, next) {
+    //add activate process
+    await postAccountSave(doc, {
+      credentialModel,
+      userModel
+    });
+    next();
+  });
+  return AccountSchema;
+};
