+++ libs/database/src/lib/services/content.service.spec.ts
@@ -38,7 +38,7 @@ import { Author, SaveContentDto, ShortPayload } from '../dtos/content.dto';
 import { EngagementDocument } from '../schemas/engagement.schema';
 import { BullModule } from '@nestjs/bull';
 import { TopicName, UserProducer } from '@castcle-api/utils/queue';
-import { UserVerified } from '../schemas/user.schema';
+import { UserVerified } from '../models';
 import { FeedItemDocument } from '../schemas/feedItem.schema';
 import { HashtagService } from './hashtag.service';
